+++ src/panels/lovelace/cards/hui-thermostat-card.ts
@@ -10,6 +10,7 @@ import "@polymer/paper-icon-button/paper-icon-button";
 
 import "../../../components/ha-card";
 import "../../../components/ha-icon";
+import "../components/hui-warning";
 
 import applyThemesOnElement from "../../../common/dom/apply_themes_on_element";
 import computeStateName from "../../../common/entity/compute_state_name";
@@ -102,16 +103,13 @@ export class HuiThermostatCard extends LitElement implements LovelaceCard {
       return html``;
     }
     const stateObj = this.hass.states[this._config.entity] as ClimateEntity;
+
     if (!stateObj) {
       return html`
-        ${this.renderStyle()}
-        <ha-card>
-          <div class="not-found">
-            Entity not available: ${this._config.entity}
-          </div>
-        </ha-card>
+        <hui-warning .entity="${this._config.entity}"></hui-warning>
       `;
     }
+
     const mode = modeIcons[stateObj.attributes.operation_mode || ""]
       ? stateObj.attributes.operation_mode!
       : "unknown-mode";
@@ -383,11 +381,6 @@ export class HuiThermostatCard extends LitElement implements LovelaceCard {
           --dry-color: #efbd07;
           --idle-color: #8a8a8a;
           --unknown-color: #bac;
-        }
-        .not-found {
-          flex: 1;
-          background-color: yellow;
-          padding: 8px;
         }
         #root {
           position: relative;
